Updated user deletion in the database
1) User deletion
If a user has tasks, the user will be set as inactive.
If the user has no tasks, the user can be deleted permanently.
-> This is needed in the scenario where a user was made by mistake.
2) Get users
Only users who are active will be shown in the UI.
3) User creation
When a user is created in the database, the created user will have the active status.

TODO: The EventLog is not completed for the new methods, because I am not sure the implementation is correct.